Variable controls

+++ src/TrailBuffer.cpp
@@ -168,13 +168,13 @@ void TrailBuffer::ZoomTrails(float zoom_factor) {
   for (size_t i = wxMax(m_trail_size / 2 + m_offset.lat - m_max_spoke_len, 0);
        i < wxMin(m_trail_size / 2 + m_offset.lat + m_max_spoke_len, m_trail_size); i++) {
     int index_i = (int((float)(i - m_trail_size / 2 + m_offset.lat) * zoom_factor)) + m_trail_size / 2 - m_offset.lat * zoom_factor;
-    if (index_i >= m_trail_size - 1) break;  // allow adding an additional pixel later
+    if (index_i >= (int)m_trail_size - 1) break;  // allow adding an additional pixel later
     if (index_i < 0) continue;
     for (size_t j = wxMax(m_trail_size / 2 + m_offset.lon - m_max_spoke_len, 0);
          j < wxMin(m_trail_size / 2 + m_offset.lon + m_max_spoke_len, m_trail_size); j++) {
       int index_j =
           (int((float)(j - m_trail_size / 2 + m_offset.lon) * zoom_factor)) + m_trail_size / 2 - m_offset.lon * zoom_factor;
-      if (index_j >= m_trail_size - 1) break;
+      if (index_j >= (int)m_trail_size - 1) break;
       if (index_j < 0) continue;
       uint8_t pixel = M_TRUE_TRAILS(i, j);
       if (pixel != 0) {  // many to one mapping, prevent overwriting trails with 0
